Query plan execution refactoring (#1843)
split query plan execution in multiple modules to make the code more
manageable
